+++ ricaun.Revit.UI/RibbonItemDataExtension.cs
@@ -90,7 +90,7 @@ namespace ricaun.Revit.UI
         }
 
         /// <summary>
-        /// Set ButtonData Image
+        /// Set ButtonData/ComboBoxData Image
         /// </summary>
         /// <typeparam name="TRibbonItem"></typeparam>
         /// <param name="ribbonItem"></param>
@@ -102,7 +102,7 @@ namespace ricaun.Revit.UI
         }
 
         /// <summary>
-        /// Set ButtonData LargeImage
+        /// Set ButtonData/ComboBoxData LargeImage
         /// </summary>
         /// <typeparam name="TRibbonItem"></typeparam>
         /// <param name="ribbonItem"></param>
@@ -114,7 +114,7 @@ namespace ricaun.Revit.UI
         }
 
         /// <summary>
-        /// Set ButtonData Image
+        /// Set ButtonData/ComboBoxData Image
         /// </summary>
         /// <typeparam name="TRibbonItem">ButtonData</typeparam>
         /// <param name="ribbonItem"></param>
@@ -122,15 +122,19 @@ namespace ricaun.Revit.UI
         /// <returns></returns>
         public static TRibbonItem SetImage<TRibbonItem>(this TRibbonItem ribbonItem, ImageSource image) where TRibbonItem : RibbonItemData
         {
-            if (image != null)
-                if (ribbonItem is ButtonData ribbonButton)
-                    ribbonButton.Image = image.GetBitmapFrame(16, (frame) => { ribbonButton.Image = frame; });
+            if (image == null) return ribbonItem;
+
+            if (ribbonItem is ButtonData ribbonButton)
+                ribbonButton.Image = image.GetBitmapFrame(16, (frame) => { ribbonButton.Image = frame; });
+
+            if (ribbonItem is ComboBoxData comboBox)
+                comboBox.Image = image.GetBitmapFrame(16, (frame) => { comboBox.Image = frame; });
 
             return ribbonItem;
         }
 
         /// <summary>
-        /// Set ButtonData LargeImage
+        /// Set ButtonData/ComboBoxData LargeImage
         /// </summary>
         /// <typeparam name="TRibbonItem">ButtonData</typeparam>
         /// <param name="ribbonItem"></param>
@@ -138,15 +142,17 @@ namespace ricaun.Revit.UI
         /// <returns></returns>
         public static TRibbonItem SetLargeImage<TRibbonItem>(this TRibbonItem ribbonItem, ImageSource largeImage) where TRibbonItem : RibbonItemData
         {
-            if (largeImage != null)
+            if (largeImage == null) return ribbonItem;
+
+            if (ribbonItem is ButtonData ribbonButton)
             {
-                if (ribbonItem is ButtonData ribbonButton)
-                {
-                    ribbonButton.LargeImage = largeImage.GetBitmapFrame(32, (frame) => { ribbonButton.LargeImage = frame; });
-                    if (ribbonButton.Image == null)
-                        ribbonButton.SetImage(ribbonButton.LargeImage);
-                }
+                ribbonButton.LargeImage = largeImage.GetBitmapFrame(32, (frame) => { ribbonButton.LargeImage = frame; });
+                if (ribbonButton.Image == null)
+                    ribbonButton.SetImage(ribbonButton.LargeImage);
             }
+
+            if (ribbonItem is ComboBoxData comboBox)
+                comboBox.SetImage(largeImage);
 
             return ribbonItem;
         }