+++ ricaun.Revit.UI/RibbonItemExtension.cs
@@ -133,7 +133,7 @@ namespace ricaun.Revit.UI
 
         #region Set RibbonButton
         /// <summary>
-        /// Set RibbonButton Image
+        /// Set RibbonButton/ComboBox Image
         /// </summary>
         /// <typeparam name="TRibbonItem"></typeparam>
         /// <param name="ribbonItem"></param>
@@ -145,7 +145,7 @@ namespace ricaun.Revit.UI
         }
 
         /// <summary>
-        /// Set RibbonButton LargeImage
+        /// Set RibbonButton/ComboBox LargeImage
         /// </summary>
         /// <typeparam name="TRibbonItem"></typeparam>
         /// <param name="ribbonItem"></param>
@@ -157,7 +157,7 @@ namespace ricaun.Revit.UI
         }
 
         /// <summary>
-        /// Set RibbonButton Image
+        /// Set RibbonButton/ComboBox Image
         /// </summary>
         /// <typeparam name="TRibbonItem">RibbonButton</typeparam>
         /// <param name="ribbonItem"></param>
@@ -165,15 +165,19 @@ namespace ricaun.Revit.UI
         /// <returns></returns>
         public static TRibbonItem SetImage<TRibbonItem>(this TRibbonItem ribbonItem, ImageSource image) where TRibbonItem : RibbonItem
         {
-            if (image != null)
-                if (ribbonItem is RibbonButton ribbonButton)
-                    ribbonButton.Image = image.GetBitmapFrame(16, (frame) => { ribbonButton.Image = frame; });
+            if (image == null) return ribbonItem;
+
+            if (ribbonItem is RibbonButton ribbonButton)
+                ribbonButton.Image = image.GetBitmapFrame(16, (frame) => { ribbonButton.Image = frame; });
+
+            if (ribbonItem is ComboBox comboBox)
+                comboBox.Image = image.GetBitmapFrame(16, (frame) => { comboBox.Image = frame; });
 
             return ribbonItem;
         }
 
         /// <summary>
-        /// Set RibbonButton LargeImage
+        /// Set RibbonButton/ComboBox LargeImage
         /// </summary>
         /// <typeparam name="TRibbonItem">RibbonButton</typeparam>
         /// <param name="ribbonItem"></param>
@@ -181,15 +185,17 @@ namespace ricaun.Revit.UI
         /// <returns></returns>
         public static TRibbonItem SetLargeImage<TRibbonItem>(this TRibbonItem ribbonItem, ImageSource largeImage) where TRibbonItem : RibbonItem
         {
-            if (largeImage != null)
+            if (largeImage == null) return ribbonItem;
+
+            if (ribbonItem is RibbonButton ribbonButton)
             {
-                if (ribbonItem is RibbonButton ribbonButton)
-                {
-                    ribbonButton.LargeImage = largeImage.GetBitmapFrame(32, (frame) => { ribbonButton.LargeImage = frame; });
-                    if (ribbonButton.Image == null)
-                        ribbonButton.SetImage(ribbonButton.LargeImage);
-                }
+                ribbonButton.LargeImage = largeImage.GetBitmapFrame(32, (frame) => { ribbonButton.LargeImage = frame; });
+                if (ribbonButton.Image == null)
+                    ribbonButton.SetImage(ribbonButton.LargeImage);
             }
+
+            if (ribbonItem is ComboBox comboBox)
+                comboBox.SetImage(largeImage);
 
             return ribbonItem;
         }