+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [0.0.8] / 2022-02-09
+- Add ComboBox Items on `GetRibbonItems`
+- Add ComboBox
 - Fix MoveRibbonPanel - Add Max Move / Add Negative
 
 ## [0.0.7] / 2022-02-08